Added Flash Scrren  and Home page
