+++ data/kannada.ts
@@ -1,0 +1,33 @@
+// export const consonantCards: FlashCardType[] = [
+//     { letter: "ಅ", word: "ಅಕ್ಕ", pronunciation: "Akka", image: Akka, translation: "Elder sister" },
+//     { letter: "ಆ", word: "ಆನೆ", pronunciation: "Āne", image: Aane, translation: "Elephant" },
+//     { letter: "ಇ", word: "ಇಲಿ", pronunciation: "Ili", image: Ili, translation: "Mouse" },
+//     { letter: "ಈ", word: "ಈಜು", pronunciation: "Īju", image: Iju, translation: "Swim" },
+//     { letter: "ಉ", word: "ಉಂಡೆ", pronunciation: "Uṇḍe", image: Unde, translation: "Laddu" },
+//     { letter: "ಊ", word: "ಊಟ", pronunciation: "Ūṭa", image: Oota, translation: "Meal" },
+//     { letter: "ಎ", word: "ಎಲೆ", pronunciation: "Ele", image: Ele, translation: "Leaf" },
+//     { letter: "ಏ", word: "ಏನು", pronunciation: "Ēnu", image: Enu, translation: "What" },
+//     { letter: "ಐ", word: "ಐದು", pronunciation: "Aidu", image: Aidu, translation: "Five" },
+//     { letter: "ಒ", word: "ಒಂಟಿ", pronunciation: "Oṇṭi", image: Oṇṭi, translation: "Alone" },
+//     { letter: "ಓ", word: "ಓದು", pronunciation: "Ōdu", image: Odu, translation: "Read" },
+//     { letter: "ಔ", word: "ಔಷಧಿ", pronunciation: "Auṣadhi", image: Aushadhi, translation: "Medicine" },
+//     { letter: "ಕ", word: "ಕಪು", pronunciation: "Kapu", image: Kapu, translation: "Monkey" },
+//     { letter: "ಖ", word: "ಖಡ್ಗ", pronunciation: "Khaḍga", image: Khadga, translation: "Sword" },
+//     { letter: "ಗ", word: "ಗೋಧಿ", pronunciation: "Gōdhi", image: Godhi, translation: "Wheat" },
+//     { letter: "ಘ", word: "ಘಂಟೆ", pronunciation: "Ghaṇṭe", image: Ghante, translation: "Bell" },
+//     { letter: "ಚ", word: "ಚೂರು", pronunciation: "Cūru", image: Churu, translation: "Small piece" },
+//     { letter: "ಜ", word: "ಜಲ", pronunciation: "Jala", image: Jala, translation: "Water" },
+//     { letter: "ಟ", word: "ಟೋಪಿ", pronunciation: "Ṭōpi", image: Topi, translation: "Hat" },
+//     { letter: "ಡ", word: "ಡಬ್ಬಿ", pronunciation: "Ḍabbi", image: Dabbi, translation: "Box" },
+//     { letter: "ತ", word: "ತಾಯಿ", pronunciation: "Tāyi", image: Taayi, translation: "Mother" },
+//     { letter: "ನ", word: "ನಮಸ್ಕಾರ", pronunciation: "Namaskāra", image: Namaskara, translation: "Greeting" },
+//     { letter: "ಪ", word: "ಪದ", pronunciation: "Pada", image: Pada, translation: "Word" },
+//     { letter: "ಮ", word: "ಮನೆ", pronunciation: "Mane", image: Mane, translation: "House" },
+//     { letter: "ಯ", word: "ಯಶಸ್ಸು", pronunciation: "Yaśassu", image: Yashassu, translation: "Success" },
+//     { letter: "ರ", word: "ರಥ", pronunciation: "Ratha", image: Ratha, translation: "Chariot" },
+//     { letter: "ಲ", word: "ಲವಣೆ", pronunciation: "Lavaṇe", image: LavaNe, translation: "Salt" },
+//     { letter: "ವ", word: "ವೃಕ್ಷ", pronunciation: "Vṛkṣa", image: Vriksha, translation: "Tree" },
+//     { letter: "ಶ", word: "ಶಕ್ತಿ", pronunciation: "Śakti", image: Shakti, translation: "Power" },
+//     { letter: "ಸ", word: "ಸೂರ್ಯ", pronunciation: "Sūrya", image: Surya, translation: "Sun" },
+//     { letter: "ಹ", word: "ಹಸು", pronunciation: "Hasu", image: Hasu, translation: "Cow" }
+//   ];
